requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
pydub
gTTS
python-dotenv
aiofiles
pytest
httpx
pytest-asyncio